project_pokdo_backend/main_site/serializers.py
from rest_framework import serializers
from .models import Listing

class ListingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Listing
        fields = ['school_name', 'about', 'location', 'admissions', 'image']