project_pokdo_backend/requirements.txt
django-environ==0.4.5
django-extensions==3.1.1
Django==3.0.8
mysqlclient==1.4.6
django-filter==2.2.0
djangorestframework==3.11.0
django-cors-headers==3.2.1